% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angleplot.R
\name{angleplot}
\alias{angleplot}
\title{Plot the two halves of the curve with touchlines and display angle}
\usage{
angleplot(
  part1 = part1,
  part2 = part2,
  test1 = test1,
  test2 = test2,
  main = kneepoint,
  angle = angle1
)
}
\arguments{
\item{part1}{Datapoints from start to kneepoint}

\item{part2}{Datapoints from kneepoint to end}

\item{test1}{Fitted line over part1}

\item{test2}{Fitted line over part2}

\item{main}{Title of the plot}

\item{angle}{Angle between test1 and test2, obtained through \code{LinesAngles} function from \code{LearnGeom}}
}
\value{
None, plot is generated using the basic R plot function.
}
\description{
Internal function. Displays basic plot of curve and touchlines created by \code{\link{leastError}}. Of greater use when refining Inflection Point
}
\seealso{
\code{\link{INFLECT}} , \code{\link{QC.to.curve}},\code{\link{leastError}}
}
