% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computemode.R
\name{computemode}
\alias{computemode}
\title{Internal - Returns the mode of a numeric vector}
\usage{
computemode(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a list with 2 numeric values specifying the mode "x" and it associated density "y"
}
\description{
This function is used internally to determine the mode of a numeric vector. Adapted from \pkg{SPADEVizR}
}
