% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FlowSOM-QC.R
\name{FlowSOMQC}
\alias{FlowSOMQC}
\title{Cluster quality control using diptest and IQR check}
\usage{
FlowSOMQC(
  FlowSOM.results,
  metaclustering,
  zeroes.in = FALSE,
  only.clustering.markers = TRUE,
  acquired_markers = NULL,
  uniform.test = "both",
  th.pvalue = 0.05,
  th.IQR = 2,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{FlowSOM.results}{A FlowSOM object with completed SOM clustering, either after full FlowSOM function or after BuildSOM}

\item{metaclustering}{Vector with metacluster codes for all SOM-clusters.}

\item{zeroes.in}{Should be values at and below \code{0} be included. Recommended default for mass cytometry data is \code{FALSE}}

\item{only.clustering.markers}{If \code{TRUE} only evaluates markers specified in FlowSOM.results$map$colsUsed}

\item{acquired_markers}{Vector of column names with marker data to be evaluated by INFLECT. Ignored if \code{only.clustering.markers == TRUE}}

\item{uniform.test}{What tests are performed per marker per cluster. Options are "both", "spread" , or "unimodality" as a string.}

\item{th.pvalue}{Threshold for rejecting Unimodality dip.test result. Default is \code{0.05}. For more information see \link[diptest]{dip.test}}

\item{th.IQR}{Threshold for rejecting marker distribution based on inter-quartile range. Default is arc-sinh transformed value of \code{2}.}

\item{verbose}{\code{logical} , default is \code{FALSE}}

\item{...}{Additional arguments to pass to this function. Not used currently.}
}
\value{
A \code{matrix} with evaluated markers in columns and clusters in rows. Each position in the matrix is \code{logical} indicating a pass or a fail.
}
\description{
Unimodality score is calculated for cluster results. Per marker per cluster \link[diptest]{dip.test} is applied and inter-quartile range is assessed.
}
\seealso{
\code{\link{INFLECT}} , \code{\link{iteration.QC}}
}
