% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFLECT3-package.R
\docType{package}
\name{INFLECT-package}
\alias{INFLECT}
\alias{INFLECT-package}
\title{INFLECT: Discover the metaclustering endpoint for FlowSOM}
\description{
Quality control and determination of optimal k for FlowSOM high-dimensional clustering data. Iteratively clusters and evaluates resulting clusters for marker unimodality. Translates this into a unimodality score and plots it into a diagnostic graph. On the fitted curve of this unimodality score INFLECT determines the inflection point where unimodality stops increasing for greater numbers of clusters.
}
\keyword{internal}
