% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INFLECT.R
\name{INFLECT}
\alias{INFLECT}
\title{Run the INFLECT computation}
\usage{
INFLECT(
  FlowSOM.results,
  set.i = c(150, 200),
  multicore = TRUE,
  cores = NULL,
  zeroes.in = FALSE,
  only.clustering.markers = TRUE,
  acquired_markers = NULL,
  basedata = "Curve",
  ggtitle = NULL,
  ...
)
}
\arguments{
\item{FlowSOM.results}{A FlowSOM object with completed SOM clustering, either after full FlowSOM function or after BuildSOM}

\item{set.i}{Vector containing either the desired iterations to be tested, or if \code{length(set.i)==2}, the point where iterations are spaced apart by 5 or 10. Intended to limit computation time}

\item{multicore}{logical, should INFLECT be run in parallel using \link[doParallel]{registerDoParallel}}

\item{cores}{If \code{multicore == TRUE}, number of cores to be used. If \code{NULL} max number of cores-1 is used}

\item{zeroes.in}{Should be values at and below \code{0} be included. Recommeded default is \code{FALSE}}

\item{only.clustering.markers}{If \code{TRUE} only evaluates markers specified in FlowSOM.results$map$colsUsed}

\item{acquired_markers}{Vector of column names with marker data to be evaluated by INFLECT. Ignored if \code{only.clustering.markers == TRUE}}

\item{basedata}{Data to be used to calculate inflection point, given as a string. Options are \code{Curve} and \code{Points}}

\item{ggtitle}{Optional title for resulting diagnostic graph. Default \code{NULL}}

\item{...}{Arguments to pass to \code{\link{FlowSOMQC}} (uniform.test, th.pvalue, th.IQR)}
}
\value{
A \linkS4class{inflect.results} object containing a dataframe with Unimodality scores, a dataframe with the fitted curve,  the \code{\link{Lfunction}} results, the diagnostic \pkg{ggplot2} plot, the \code{list} of metaclusterings performed during \code{\link{iteration.metacluster}}, and a \code{list} of accuracy matrices generated by \code{\link{FlowSOMQC}}
Running the individual function \code{\link{iteration.metacluster}} and \code{\link{QC.to.curve}} might provide more options and flexibility.
}
\description{
Wrapper function combining all aspects of INFLECT, iterative clustering, unimodality testing and inflection point determination.
Default options include fewer calculations at higher cluster numbers, zeroes are removed and inflection point is calculated on the fitted curve.
}
\examples{

# Read in FlowSOM object from file. Downsampled clustering result of Levine32 dataset clustering.
# SOM-clustered to 375 clusters.
flowsom <- system.file("extdata", "Levine32sample.Rdata", package="INFLECT")
load(flowsom)
set.i<- c( seq(5, 50), seq(55,100,5), seq(110,200, 10))
inflect.results<- INFLECT(FlowSOM.results= dataset, set.i= set.i, multicore=TRUE, cores=4, zeroes.in=FALSE)

# Display diagnostic graph
inflect.results$ggplot

}
\seealso{
\code{\link{iteration.metacluster}}, \code{\link{iteration.QC}}, \code{\link{FlowSOMQC}}, \code{\link{QC.to.curve}}, \code{\link{leastError}}, \code{\link{Lfunction}}, \code{\link{marker.performance}}
}
