% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflect-results-class.R
\docType{class}
\name{inflect.results-class}
\alias{inflect.results-class}
\alias{inflect.results}
\title{inflect.results class definition}
\description{
The inflect.results object is an S4 object containing the Inflection Point, diagnostic curve and associated datapoints for use of customization of graphs.
}
\details{
The \code{collection.U} dataframe contains the Unimodality scores for each \code{i} in \code{set.i}.
The \code{fittedcurve} dataframe contains the coordinates of the curve fitted to \code{collection.U}.
The \code{lfunction} is a \code{list} with results from \code{\link{Lfunction}}, denoting the Inflection Point and touchline range.
The \code{ggplot} is the \code{ggplot2} object of the diagnostic graph.
}
\section{Slots}{

\describe{
\item{\code{collection.U}}{dataframe containing the Unimodality scores for each \code{i} in \code{set.i}.}

\item{\code{fittedcurve}}{dataframe containing the coordinates of the curve fitted to \code{collection.U}.}

\item{\code{lfunction}}{\code{list} with results from \code{\link{Lfunction}}, denoting the Inflection Point and touchline range.}

\item{\code{ggplot}}{the \code{ggplot2} object of the diagnostic graph}

\item{\code{metaclustering.list}}{list containing the iterative metaclustering results according to \code{set.i}}

\item{\code{accuracy.sets}}{list containing \link{FlowSOMQC} results per metaclustering}
}}

