% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration-metacluster.R
\name{iteration.metacluster}
\alias{iteration.metacluster}
\title{Run iterative metaclustering on SOM-clustered FlowSOM object}
\usage{
iteration.metacluster(FlowSOM.results, set.i, multicore = TRUE, cores = NULL)
}
\arguments{
\item{FlowSOM.results}{A FlowSOM object with completed SOM clustering, either after full FlowSOM function or after BuildSOM}

\item{set.i}{Vector containing either the desired iterations to be tested}

\item{multicore}{logical, should INFLECT be run in parallel using \link[doParallel]{registerDoParallel}}

\item{cores}{If \code{multicore == TRUE}, number of cores to be used. If \code{NULL} max number of cores-1 is used}
}
\value{
metaclustering.list A \code{list} of \code{arrays} with metacluster-codes for the SOM-clusters within the FlowSOM.results object.
}
\description{
Metaclustering runs are determined for all cluster numbers in set.i . Hierarchical clustering is used, optionally run in parallel
}
\seealso{
\code{\link{INFLECT}}
}
