% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iteration-QC.R
\name{iteration.QC}
\alias{iteration.QC}
\title{Run unimodality quality control on metaclusterings in parallel}
\usage{
iteration.QC(
  FlowSOM.results,
  metaclustering.list,
  set.i,
  multicore = TRUE,
  cores = NULL,
  zeroes.in = FALSE,
  only.clustering.markers = TRUE,
  acquired_markers = NULL,
  uniform.test = "both",
  th.pvalue = 0.05,
  th.IQR = 2,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{FlowSOM.results}{A FlowSOM object with completed SOM clustering, either after full FlowSOM function or after BuildSOM}

\item{metaclustering.list}{List of metaclustering results, as generated by \link{iteration.metacluster}}

\item{set.i}{Vector containing either the desired iterations to be tested}

\item{multicore}{logical, should INFLECT be run in parallel using \link[doParallel]{registerDoParallel}. Default is \code{TRUE}}

\item{cores}{If \code{multicore == TRUE}, number of cores to be used. If \code{NULL} max number of cores-1 is used.}

\item{zeroes.in}{Should be values at and below \code{0} be included. Recommeded default for mass cytometry data is \code{FALSE}}

\item{only.clustering.markers}{If \code{TRUE} only evaluates markers specified in FlowSOM.results$map$colsUsed}

\item{acquired_markers}{Vector of column names with marker data to be evaluated by INFLECT. Ignored if \code{only.clustering.markers == TRUE}}

\item{uniform.test}{What tests are performed per marker per cluster. Options are "both", "spread" , or "unimodality" as a string.}

\item{th.pvalue}{Threshold for rejecting Unimodality dip.test result. Default is \code{0.05}. For more information see \link[diptest]{dip.test}}

\item{th.IQR}{Threshold for rejecting marker distribution based on inter-quartile range. Default is arc-sinh transformed value of \code{2}.}

\item{verbose}{\code{logical} , default is \code{FALSE}}

\item{...}{Additional arguments to pass to this function. Not used currently.}
}
\value{
A \code{list} with 2 items. The first is the unimodality scores stored in U.set. Second is a \code{list} with a \code{logical} matrix for each metaclustering result.
}
\description{
Unimodality score is calculated for all calculated metaclusterings.
}
\seealso{
\code{\link{INFLECT}}, \code{\link{iteration.metacluster}}, \code{\link{FlowSOMQC}}
}
