% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastError.R
\name{leastError}
\alias{leastError}
\title{Determine kneepoint based on minimizing fitting errors}
\usage{
leastError(dataframe)
}
\arguments{
\item{dataframe}{Dataframe with set.i and corresponding unimodality scores, resulting from \code{\link{iteration.QC}}}
}
\value{
Integer denoting the row of dataframe whose coordinate provides the kneepoint
}
\description{
Evaluates the errors for two fitted lines on both parts of a collection of points. Function calculates the errors for a split on each element and selects the one with least amount of error. Internal function of \code{\link{Lfunction}}
}
\seealso{
\code{\link{INFLECT}} , \code{\link{QC.to.curve}},\code{\link{Lfunction}}
}
