% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lfunction.R
\name{Lfunction}
\alias{Lfunction}
\title{Lfunction for determing Inflection Point}
\usage{
Lfunction(totaldataframe, cutoff = 1000)
}
\arguments{
\item{totaldataframe}{Dataframe with set.i and corresponding unimodality scores, resulting from \code{\link{iteration.QC}}. This can be the calculcated points collection.U or the fitted curve.}

\item{cutoff}{Integer. Initial Inflection Point calculated on the entire curve is multiplied by \code{cutoff} to determine the new range of the curve that is used as input for \code{\link{leastError}}}
}
\value{
\code{list} with 3 items: Inflection Point, the endpoint of the second touchline, and the angle between the two touchlines
}
\description{
Wrapper function for \code{\link{leastError}} with an option for refining the Inflection Point, by limiting the range of of points to evaluate for fitting line errors. Default is to not use the refinement.
}
\seealso{
\code{\link{INFLECT}} , \code{\link{QC.to.curve}},\code{\link{leastError}}
}
