% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-markerperformance.R
\name{marker.performance}
\alias{marker.performance}
\title{Plot marker performance across metaclustering results}
\usage{
marker.performance(inflect.results, ggtitle = NULL, markers)
}
\arguments{
\item{inflect.results}{A inflect.results object resulting from \code{\link{INFLECT}}.}

\item{ggtitle}{Optional. Character string to plot as title.}

\item{markers}{Which markers should be included in the plot? A vector of strings (must match FlowSOM.results$prettyColnames). If \code{NULL}, all clustering.markers are displayed.}
}
\value{
\code{list} with 2 items. First is the melted dataframe with the marker performance (\\%) per marker per metaclustering. Second is a ggplot object, a boxplot with marker
performance on the y-axis, markers on the x-axis and the color scale denoting the amount of metaclusters evaluated.
}
\description{
Plots the marker performance across metaclusterings in a boxplot. Performance score is the percentage of clusters
where a marker passed the \link[diptest]{dip.test} and IQR test per metaclustering. Information is displayed in a boxplot,
color gradient denotes \code{set.i} .
}
\examples{
# Read in FlowSOM object from file. Downsampled clustering result of Levine32 dataset clustering.
# SOM-clustered to 375 clusters.
flowsom <- system.file("extdata", "Levine32sample.Rdata", package="INFLECT")
load(flowsom)
inflect.results<- INFLECT(FlowSOM.results= dataset, set.i= c(150,200), cores= 4, zeroes.in=FALSE)

# Display diagnostic graph
inflect.results$ggplot

# Now check marker performance for all markers
marker.performance(inflect.results, ggtitle= "Levine32sample", markers=NULL)


}
