% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaClustering-hclust.R
\name{metaClusteringhclust}
\alias{metaClusteringhclust}
\title{Metaclustering function based on hierarchical clustering}
\usage{
metaClusteringhclust(data, nClus)
}
\arguments{
\item{data}{Matrix with the median values for each clustering parameter for all SOM clusters.}

\item{nClus}{Amount of metaclusters to be obtained.}
}
\value{
A \code{scalar} of \code{length(SOMclusters)} with metacluster identity.
}
\description{
Orphaned function from FlowSOM, imported here.
}
\seealso{
\code{\link{INFLECT}}
}
